+++ .npmignore
@@ -1,0 +1,3 @@
+.babelrc
+webpack.config.js
+/src/
