+++ package.json
@@ -13,7 +13,8 @@
     "node": ">=4"
   },
   "scripts": {
-    "test": "xo && ava"
+    "test": "xo && ava",
+    "prepublish": "webpack --progress --colors --display-error-details --config webpack.config.js"
   },
   "files": [
     "index.js"
@@ -37,6 +38,12 @@
   },
   "devDependencies": {
     "ava": "*",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.24.1",
+    "uglify-es": "git://github.com/mishoo/UglifyJS2.git#harmony",
+    "uglifyjs-webpack-plugin": "^0.4.3",
+    "webpack": "^2.5.1",
     "xo": "*"
   }
 }
