+++ webpack.config.js
@@ -1,0 +1,23 @@
+const path = require('path');
+const webpack = require('webpack');
+const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
+
+module.exports = {
+  entry: ['./src/index.js'],
+  output: {
+    path: path.join(__dirname, './lib'),
+    filename: 'bundle.js',
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js$/,
+        exclude: /(node_modules)/,
+        use: ['babel-loader']
+      }
+    ]
+  },
+  plugins: [
+    new UglifyJSPlugin()
+  ],
+};
